vtab/pyproject.toml
[tool.poetry]
# TODO: setme
name = "otdd_map"
version = "0.0.1"
description = "Monge map for OTDD"
authors = ["sbyebss"]
packages = [
    { include="src" },
]


[tool.poetry.dependencies]
python = ">=3.7, <3.10"
torch = "^1.9.0"
torchvision = "^0.10.0"
pytorch-lightning = "^1.3.8"
torchmetrics = "^0.6.0"
hydra-core = "^1.1.0"
hydra-colorlog = "^1.1.0"
hydra-joblib-launcher = "^1.1.5"
# hydra-optuna-sweeper = "1.1.0", not compatible
wandb = "^0.10.30"
rich = "^10.6.0"
kornia = "^0.5.6"
scikit-learn = "^0.24.2"
scikit-image = "^0.18.0"
einops = "0.3.0"
sh = "^1.14.2"
seaborn = "^0.11.2"
python-dotenv = "^0.19.0"
ipdb = "^0.13.0"
setuptools = "^59.6.0"
torch-ema = "^0.3"
torch-fidelity = "^0.3.0"
jupyter = "^1.0.0"
ipywidgets = "^7.6.5"
widgetsnbextension = "^3.5.2"
pandas-profiling = "^3.1.0"
geomloss = "^0.2.5"
python-ternary = "^1.0.8"
qpsolvers = "^2.2.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.13.0"
jupyterlab = "^3.0.1"
pytest = "^6.2.1"
pylint = "^2.9.5"
jammy = "0.0.6"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools ~= 58.0"]
build-backend = "poetry.core.masonry.api"
